fix(multiplexer): absolute pane containers + review fixes + cleanups
Black-pane root cause: pane containers were Relative flex children, so
sibling containers flex-shared the window container's area and the
absolute pane rects (window-container coordinates) resolved against a
displaced origin — a split's new pane rendered entirely off-screen. Pin
every pane container (bootstrap, split, new-window) to an absolute
full-size node at the container origin, with regression tests.

Code-review fixes (15 confirmed findings): Enter takes the advertised
[y/N] default; prompts ignore chord-modified keys, capture their target
window and keyboard cursor at open time (no arming-frame swallow),
accept Key::Space, and receive IME commits; divider hit-tests subtract
the WindowContainer origin, drags bind to their window and axis, and a
grab-zone press no longer double-dispatches into panes; directional
select lands on the adjacent pane, not the farthest; resize dies at the
nearest clamped split; a rejected split keeps zoom; a split moves
KeyboardFocused in the same flush; window-bar clicks keep u32 indices;
pane styles resync on window switch.

Simplify pass: ModalKeys + shared bottom-bar helpers in modal.rs;
DividerDrag::claims as the single press-claim point (now also guarding
the webview router); spawn_pane_container/insert_spawned_pane
deduplicate the pane spawn ritual; cell_dims / PaneRect::from_size /
rect_logical_vals consolidate geometry math; divider hover gated off
idle frames, handles reconciled in place, one layout clone per
drag/resize run; window bar skips identical rebuilds; PanePlugin merges
the three per-system plugins; assorted dead-guard and doc cleanups.

+++ src/input/shortcuts/multiplexer.rs
@@ -119,7 +119,7 @@ fn apply_multiplexer_shortcuts(
             Shortcut::SelectWindow(index) => {
                 requests
                     .select_window
-                    .write(SelectWindowRequest(WindowSelect::Index(index)));
+                    .write(SelectWindowRequest(WindowSelect::Index(u32::from(index))));
             }
             Shortcut::RenameWindow => {
                 requests.rename_window.write(RenameWindowRequest);
